Better docker ports (#81)
* Do not claim the well-known ports

This refines the docker configuration to not attempt to claim the
well-known ports for postgres and redis.  This allows multiple different
docker application stack to be running, without clashing on related
services.

The .env configuration will read the randomly assigned port for redis
and the postgres database, and assign the environment variables when the
application boots.  It also provides a localhost / well-known port
fallback if the docker compose service is not running -- which allows
for fully-local development, if docker isn't your cup of tea

* allow binding to a port other than 3000