Add models
This includes work for adding a few new models to our app.

GradeBook => Represents a grade book for a specific Classroom and Quarter. A Gradebook will have three different statuses: draft, verified, and completed.

A draft Gradebook will represent a WIP. A "verified" Gradebook will represent a completed and verified gradebook that is ready for "scoring"/point allocation.

A "completed" gradebook represents a gradebook that has been filled out, verified, and has had all the points/money distributed to students.

Quarter => Represents a moment in time for a given School Year. Quarters will act as a "bucket" of sorts to connect a Classroom to a specific GradeBook

GradeEntry => Represents the actual grade and attendance entries for a single student, within a gradebook.