Add temporary UI for Grade Books
This adds a very basic, temporary UI to the Classroom show page. We need a way to view Grade Books and until the tabbed table views are implemented here, I figured a simple list with links would suffice. Once the tabbed tables are implemented we can remove this list of links.