Update styling
This just updates some of our styling so that our table views within the app actually stretch across more of the view port. Previously, things were only using about 60% of the available space.

I'm not a CSS wizard though and am unsure if we WANT to stick with these changes, but they feel better to me (at least in the interim)