Implement autosave, update tests
This introduces a basic autosave feature to the grade book page, additionally, I've refactored some of the specs so they are a little simpler.

Note: The initial version of this feature had the manual save grades button hidden, but I decided it might make more sense to allow the Teacher/Admin the ability to both manually save AND have an autosave.

Things felt a bit weird with just autosaving only.