Issue-660: create next and previous for year model
We'd like for a quarter to be smarter and understand previous and next to sequence
across years. To support that, we need to make years smarter in the same way. That way
we can ensure that traversing relationships from quarter to school_year to year -- just
works.

See Issue-660 for architectural design analysis.